requirements.txt
transformers>=4.33.0
datasets>=2.14.0
accelerate>=0.21.0
PyYAML>=6.0
scikit-learn>=1.0.0
torch>=2.0.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
